change: Database and WritableDatabase - New can be both sync and async. On async it needs a callback and doesn't emit
