+++ wscript
@@ -24,7 +24,7 @@ def build(bld):
   obj = bld.new_task_gen("cxx", "shlib", "node_addon")
   obj.cxxflags = ["-g", "-D_FILE_OFFSET_BITS=64", "-D_LARGEFILE_SOURCE", "-Wall"]
   obj.target = "xapian-binding"
-  obj.source = [ "src/enquire.cc", "src/binding.cc" ]
+  obj.source = [ "src/enquire.cc", "src/database.cc", "src/writabledatabase.cc", "src/binding.cc" ]
   #obj.lib = "xapian"
   obj.uselib = "XAPIAN PROFILER"
